adding firebase

+++ public/js/app.js
@@ -1,12 +1,13 @@
 let toss_val;
-let username = "Joemama"
+let username = "Joemama";
+
+var database = firebase.database();
+
 jQuery(document).ready(function ($) {
   $("#coin").on("click", function () {
     if (document.getElementById("paanch").checked) {
       toss_val = "Paanch";
-    }
-    else
-      toss_val = "Chaand";
+    } else toss_val = "Chaand";
     var flipResult = Math.random(0, 1);
     console.log(flipResult);
     $("#coin").removeClass();
@@ -15,12 +16,14 @@ jQuery(document).ready(function ($) {
         $("#coin").addClass("heads");
         setTimeout(() => {
           if (toss_val === "Chaand") {
-            document.getElementsByClassName("selectInnings")[0].style.display = "block";
+            document.getElementsByClassName("selectInnings")[0].style.display =
+              "block";
             document.getElementById("result").innerHTML = "You Won the toss";
-          }
-          else {
-            document.getElementsByClassName("selectInnings")[0].style.display = "none";
-            document.getElementById("result").innerHTML = "You Lost the toss<br> Other Player is batting first";
+          } else {
+            document.getElementsByClassName("selectInnings")[0].style.display =
+              "none";
+            document.getElementById("result").innerHTML =
+              "You Lost the toss<br> Other Player is batting first";
             localStorage.setItem("player", 1);
             document.getElementById("startGame").style.opacity = 1;
             document.getElementById("startGame").style.pointerEvents = "auto";
@@ -31,12 +34,14 @@ jQuery(document).ready(function ($) {
         $("#coin").addClass("tails");
         setTimeout(() => {
           if (toss_val === "Paanch") {
-            document.getElementsByClassName("selectInnings")[0].style.display = "block";
+            document.getElementsByClassName("selectInnings")[0].style.display =
+              "block";
             document.getElementById("result").innerHTML = "You Won the toss";
-          }
-          else {
-            document.getElementsByClassName("selectInnings")[0].style.display = "none";
-            document.getElementById("result").innerHTML = "You Lost the toss<br> Computer is batting first";
+          } else {
+            document.getElementsByClassName("selectInnings")[0].style.display =
+              "none";
+            document.getElementById("result").innerHTML =
+              "You Lost the toss<br> Computer is batting first";
             localStorage.setItem("player", 1);
             document.getElementById("startGame").style.opacity = 1;
             document.getElementById("startGame").style.pointerEvents = "auto";
@@ -51,19 +56,17 @@ jQuery(document).ready(function ($) {
 function enableToss() {
   document.getElementById("coin").style.display = "block";
   let old = document.getElementById("coin").style.animation;
-  document.getElementById("coin").style.animation = "scale-up-center 0.4s cubic-bezier(0.390, 0.575, 0.565, 1.000) both";
+  document.getElementById("coin").style.animation =
+    "scale-up-center 0.4s cubic-bezier(0.390, 0.575, 0.565, 1.000) both";
   setTimeout(() => {
     document.getElementById("coin").style.animation = old;
   }, 1000);
-};
-
+}
 
 function selectToss() {
   if (document.getElementById("bat").checked) {
     localStorage.setItem("player", 2);
-  }
-  else
-    localStorage.setItem("player", 1);
+  } else localStorage.setItem("player", 1);
   document.getElementById("startGame").style.opacity = 1;
   document.getElementById("startGame").style.pointerEvents = "auto";
 }
@@ -74,13 +77,31 @@ function hideMe() {
   }, 800);
 }
 
-
 function addUser() {
-  if (document.getElementById("username").value !== '')
-    username = document.getElementById("username").value
-  $(document).ready(function () {
-    $("#usernameModal").modal('hide');
-  });
-  localStorage.setItem("username", username)
+  if (document.getElementById("username").value !== "")
+    username = document.getElementById("username").value;
+  if (!isUserExists) {
+    $(document).ready(function () {
+      $("#usernameModal").modal("hide");
+    });
+    localStorage.setItem("username", username);
+  } else {
+    $("#dbMessage").html(
+      "Username " + username + " already exists, try another"
+    );
+    $("#messageALert").addClass("show");
+  }
 }
 
+function isUserExists(uid) {
+  database()
+    .ref("/users/" + uid)
+    .once("value")
+    .then((snapshot) => {
+      let username = (snapshot.val() && snapshot.val().username) || null;
+      if (username != null) {
+        return true;
+      }
+      return false;
+    });
+}
